graphic fixes, wraps scroll to header function
